+++ src/cli/cmd/iexec-orderbook.js
@@ -38,6 +38,9 @@ orderbookApp
   .option(...option.includeDatasetSpecific())
   .option(...option.includeWorkerpoolSpecific())
   .option(...option.includeRequesterSpecific())
+  .option(...option.isDatasetStrict())
+  .option(...option.isWorkerpoolStrict())
+  .option(...option.isRequesterStrict())
   .description(desc.showObj('app orderbook', 'marketplace'))
   .action(async (app, opts) => {
     await checkUpdate(opts);
@@ -55,6 +58,9 @@ orderbookApp
         maxTag,
         minVolume,
         raw,
+        isDatasetStrict,
+        isWorkerpoolStrict,
+        isRequesterStrict,
       } = opts;
 
       const request = fetchAppOrderbook(
@@ -68,6 +74,9 @@ orderbookApp
           minTag: tag !== undefined ? tag : requireTag,
           maxTag: tag !== undefined ? tag : maxTag,
           minVolume,
+          isDatasetStrict,
+          isWorkerpoolStrict,
+          isRequesterStrict,
         },
       );
       const fetchMessage = info.showing(objName);
@@ -144,6 +153,9 @@ orderbookDataset
   .option(...option.includeAppSpecific())
   .option(...option.includeWorkerpoolSpecific())
   .option(...option.includeRequesterSpecific())
+  .option(...option.isAppStrict())
+  .option(...option.isRequesterStrict())
+  .option(...option.isWorkerpoolStrict())
   .description(desc.showObj('dataset orderbook', 'marketplace'))
   .action(async (dataset, opts) => {
     await checkUpdate(opts);
@@ -161,6 +173,9 @@ orderbookDataset
         maxTag,
         requireTag,
         raw,
+        isAppStrict,
+        isRequesterStrict,
+        isWorkerpoolStrict,
       } = opts;
 
       const request = fetchDatasetOrderbook(
@@ -174,6 +189,9 @@ orderbookDataset
           minTag: tag !== undefined ? tag : requireTag,
           maxTag: tag !== undefined ? tag : maxTag,
           minVolume,
+          isAppStrict,
+          isRequesterStrict,
+          isWorkerpoolStrict,
         },
       );
       const fetchMessage = info.showing(objName);
@@ -252,6 +270,9 @@ orderbookWorkerpool
   .option(...option.includeAppSpecific())
   .option(...option.includeDatasetSpecific())
   .option(...option.includeRequesterSpecific())
+  .option(...option.isDatasetStrict())
+  .option(...option.isAppStrict())
+  .option(...option.isRequesterStrict())
   .description(desc.showObj('workerpools orderbook', 'marketplace'))
   .action(async (workerpool, opts) => {
     await checkUpdate(opts);
@@ -271,6 +292,9 @@ orderbookWorkerpool
         minVolume,
         minTrust,
         raw,
+        isAppStrict,
+        isRequesterStrict,
+        isDatasetStrict,
       } = opts;
 
       const request = fetchWorkerpoolOrderbook(
@@ -286,6 +310,9 @@ orderbookWorkerpool
           app,
           dataset,
           requester,
+          isAppStrict,
+          isRequesterStrict,
+          isDatasetStrict,
         },
       );
       const fetchMessage = info.showing(objName);
@@ -372,6 +399,7 @@ orderbookRequester
   .option(...option.filterDatasetSpecific())
   .option(...option.filterBeneficiarySpecific())
   .option(...option.includeWorkerpoolSpecific())
+  .option(...option.isWorkerpoolStrict())
   .description(desc.showObj('requesters orderbook', 'marketplace'))
   .action(async (address, opts) => {
     await checkUpdate(opts);
@@ -392,6 +420,7 @@ orderbookRequester
         maxTrust,
         minVolume,
         raw,
+        isWorkerpoolStrict,
       } = opts;
 
       const request = fetchRequestOrderbook(
@@ -408,6 +437,7 @@ orderbookRequester
           dataset,
           workerpool,
           beneficiary,
+          isWorkerpoolStrict,
         },
       );
       const fetchMessage = info.showing(objName);