+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [8.6.0] 2024-03-04
+
+### Added
+
+- strict mode `isRequesterStrict`, `isAppStrict`, `isDatasetStrict`, `isWorkerpoolStrict` in corresponding orderbook methods `fetchRequestOrderbook()`, `fetchAppOrderbook()`, `fetchDatasetOrderbook()`, `fetchWorkerpoolOrderbook()`, defaults to false
+
+### Changed
+
+- remove ipfs initialization preflight checks on request orders
+- Typescript fixes
+
 ## [8.5.2] 2024-01-24
 
 ### Added